+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v0.9.2
+
+- support for case-insensitive Unicode code point searches
+- refactored command string logic with new commandlines dependency
+- removed Naked framework dependency
+
 ## v0.9.1
 
 - updated in-application help documentation with information about primary commands